admin: Split MSW handlers into separate files
Extract each handler domain into its own file under
.storybook/mocks/handlers/ so each concern is isolated and easy to find.
handlers.ts becomes a thin aggregator that re-exports the Manufacturer and
Product types (keeping existing story imports unchanged) and assembles the
handlers array.
