+++ packages/admin/admin/.storybook/mocks/handlers/users.ts
@@ -1,0 +1,47 @@
+import { graphql, HttpResponse } from "msw";
+
+const users = [
+    { id: 1, name: "Leanne Graham", username: "Bret", email: "Sincere@april.biz" },
+    { id: 2, name: "Ervin Howell", username: "Antonette", email: "Shanna@melissa.tv" },
+    { id: 3, name: "Clementine Bauch", username: "Samantha", email: "Nathan@yesenia.net" },
+    { id: 4, name: "Patricia Lebsack", username: "Karianne", email: "Julianne.OConner@kory.org" },
+    { id: 5, name: "Chelsey Dietrich", username: "Kamren", email: "Lucio_Hettinger@annie.ca" },
+    { id: 6, name: "Mrs. Dennis Schulist", username: "Leopoldo_Corkery", email: "Karley_Dach@jasper.info" },
+    { id: 7, name: "Kurtis Weissnat", username: "Elwyn.Skiles", email: "Telly.Hoeger@billy.biz" },
+    { id: 8, name: "Nicholas Runolfsdottir V", username: "Maxime_Nienow", email: "Sherwood@rosamond.me" },
+    { id: 9, name: "Glenna Reichert", username: "Delphine", email: "Chaim_McDermott@dana.io" },
+    { id: 10, name: "Clementina DuBuque", username: "Moriah.Stanton", email: "Rey.Padberg@karina.biz" },
+];
+
+export type User = (typeof users)[number];
+
+export { users };
+
+export const usersQueryHandler = graphql.query<{ users: User[] }, { query?: string; sort?: string; order?: string }>("users", ({ variables }) => {
+    const query = variables.query?.toLowerCase() ?? "";
+    const sort = variables.sort;
+    const order = variables.order?.toUpperCase() ?? "ASC";
+
+    let result = query ? users.filter((user) => user.name.toLowerCase().includes(query)) : [...users];
+
+    if (sort) {
+        result = result.sort((a, b) => {
+            const aVal = a[sort as keyof User] as string | number;
+            const bVal = b[sort as keyof User] as string | number;
+            const cmp = aVal < bVal ? -1 : aVal > bVal ? 1 : 0;
+            return order === "DESC" ? -cmp : cmp;
+        });
+    }
+
+    return HttpResponse.json({
+        data: { users: result },
+    });
+});
+
+export const userQueryHandler = graphql.query<{ user: User | null }, { id: number }>("user", ({ variables }) => {
+    return HttpResponse.json({
+        data: {
+            user: users.find((user) => user.id === variables.id) ?? null,
+        },
+    });
+});
